requirements.txt
telethon>=1.25.0
python-dotenv>=0.19.0
requests>=2.28.0
flask>=2.3.0
